requirements.txt
asgiref==3.6.0
Django==4.1.9
django-cors-headers==4.0.0
djangorestframework==3.14.0
dnspython==2.3.0
gunicorn==20.1.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.2.4
whitenoise==6.4.0
